+++ bindings/node/lib/bindings/tokenizer.test.ts
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
 /* eslint-disable @typescript-eslint/no-empty-function */
 
 import { promisify } from "util";
@@ -112,7 +113,7 @@ describe("Tokenizer", () => {
         [2, 6],
         [6, 8],
         [8, 12],
-        [12, 16]
+        [0, 4]
       ]);
       expect(encoding.getOverflowing()).toEqual([]);
       expect(encoding.getSpecialTokensMask()).toEqual([0, 0, 0, 0, 0]);